[IZPACK-986] - DataValidator, CompilerConfig - status attribute casing
The schema indicated that a status is in lower case, but the parser
demanded upper-case because it used valueOf(String) instead of the
already existing getStatusFromAttribute(String). The latter method uses
a map to lookup the values. Since the map is populated with lower-case
keys, I switched the CompilerConfig lookup to use
DataValidator.getStatusFromAttribute(String) and in turn asserted that
the that the latter method should always transform the attribute value
to lower-case prior to looking it up in the map.